Intermediate interface for configurator-based factories.
To be used by detectors and trackers.
